requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==2.1.0
dj-database-url==2.1.0
Django==4.2.4
django-environ==0.10.0
gunicorn==21.2.0
idna==3.3
packaging==23.1
Pillow==9.3.0
python-decouple==3.6
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==1.26.11
whitenoise==6.5.0
google-cloud-firestore==2.14.0
